+++ web/package.json
@@ -17,6 +17,7 @@
     "nth-check": "^2.1.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-icons": "^4.4.0",
     "react-relay": "^14.1.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
