+++ web/src/pages/Request.tsx
@@ -1,5 +1,6 @@
 // (c) Tecnologico de Monterrey 2022, rights reserved.
-import Request_Card from "../components/Request_Card";
+import Document_Review from "../components/Document_Review";
+import Request_Detail from "../components/Request_Card_Detail";
 import Foot from "./includes/Footer";
 import NavBar from "./includes/NavBar";
 
@@ -7,16 +8,17 @@ const Request = (): JSX.Element => {
     return (
         <div>
             <NavBar />
-            <div className="h-auto w-full flex gap-4 pt-16 px-24">
+            <div className="w-full flex gap-4 pt-16 px-24">
                 <div className="w-2/5 space-y-4">
-                    <Request_Card />                    
+                    <Request_Detail />
                 </div>
-                <div className="pt-6">
+                <div className="pt-6 space-y-2">
                     <h2 className="text-lg text-[#396FB1] font-medium">Documentos</h2>
-                    <p>Mod_Documento1</p>
-                    <p>Mod_Documento2</p>
-                    <p>Mod_Documento3</p>
-                    <p>Mod_Documento4</p>
+                    <div className="divide-y divide-solid space-y-4">
+                        <Document_Review />
+                        <Document_Review />
+                        <Document_Review />
+                    </div>
                     <p>Botones</p>
                 </div>
             </div>
